+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Next Release
 * Updated the Amazon S3 PostObject class to work with CNAME style buckets.
 * Added the us-west-2, eu-west-1, ap-southeast-2, and ap-northeast-1 regions to the AWS Data Pipeline client.
 * Added support for SerialNumber and TokenCode to the AssumeRole operation of the IAM Security Token Service (STS).
+* Added support for RequestInterval and FailureThreshold to the Amazon Route53 client.
 
 2.5.2 (2014-01-29)
 ------------------